requirements.txt
slack-sdk>=3.23.0
slack-bolt>=1.18.0
python-dotenv>=1.0.0
requests>=2.31.0
atlassian-python-api>=3.41.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiohttp>=3.9.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
joblib>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
imbalanced-learn>=0.11.0
